+++ chess/__init__.py
@@ -1,1 +1,4 @@
-__all__ = ["piece", "tile", "moves", "board"]
+"""
+Abstract class for n-dimensinoal chess, also available with a GUI using tkinter.
+Regular 2D-chess, and 4D chess are both available as examples
+"""
