+++ chess/assets/piece.py
@@ -1,6 +1,6 @@
 ### Chess pieces
 
-legal_names = ["Pawn", "Knight", "Rook", "Bishop", "Queen", "King", "Trebuchet", "Superqueen"]
+LEGAL_NAMES = ["Pawn", "Knight", "Rook", "Bishop", "Queen", "King", "Trebuchet", "Superqueen"]
 
 
 class Piece:
@@ -45,7 +45,7 @@ class Piece:
             IndexError: If the piece is invalid, either from invalid value or color
         """
 
-        names = [None] + legal_names
+        names = [None] + LEGAL_NAMES
         value = value.capitalize()
 
         if value not in names:
@@ -68,9 +68,9 @@ class Piece:
         Raises:
             IndexError: If the piece is invalid, either from invalid value or color
         """
-        names = [None] + legal_names
+        names = [None] + LEGAL_NAMES
 
-        if value > len(legal_names) - 1 or value < 0:
+        if value > len(LEGAL_NAMES) - 1 or value < 0:
             if value == 69:
                 self._val = "Superqueen"
             else:
@@ -104,17 +104,17 @@ class Piece:
     def __str__(self):
         if self._val is None:
             return "  "
-        elif legal_names.index(self._val) == 0:
+        elif LEGAL_NAMES.index(self._val) == 0:
             p = "p"
-        elif legal_names.index(self._val) == 1:
+        elif LEGAL_NAMES.index(self._val) == 1:
             p = "k"
-        elif legal_names.index(self._val) == 2:
+        elif LEGAL_NAMES.index(self._val) == 2:
             p = "r"
-        elif legal_names.index(self._val) == 3:
+        elif LEGAL_NAMES.index(self._val) == 3:
             p = "b"
-        elif legal_names.index(self._val) == 4:
+        elif LEGAL_NAMES.index(self._val) == 4:
             p = "Q"
-        elif legal_names.index(self._val) == 5:
+        elif LEGAL_NAMES.index(self._val) == 5:
             p = "K"
 
         return f"{'w' if self._color else 'b'}{p}"
